refactor(domain): user 패키지명을 domain으로 수정
기능에 부합하는 이름으로 변경
