+++ src/main/java/lotto/domain/Purchase.java
@@ -1,4 +1,4 @@
-package lotto.user;
+package lotto.domain;
 
 import lotto.util.Constant;
 import lotto.util.ErrorHandler;
